Refactor timeframe validation into own class
